requirement.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
Markdown==3.6
pillow==10.4.0
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
